+++ iac-extensions/terraform/src/main/java/org/sonar/iac/terraform/checks/UnencryptedCloudServicesCheck.java
@@ -21,17 +21,26 @@ package org.sonar.iac.terraform.checks;
 
 import org.sonar.check.Rule;
 import org.sonar.iac.common.api.checks.IacCheck;
-import org.sonar.iac.common.api.checks.InitContext;
-import org.sonar.iac.terraform.checks.azure.AzureUnencryptedCloudServicesCheckPart;
 
 @Rule(key = "S6388")
-public class UnencryptedCloudServicesCheck implements IacCheck
+public class UnencryptedCloudServicesCheck extends ResourceVisitor implements IacCheck
 {
-  public static final String MESSAGE_MAKE_SURE = "Make sure using unencrypted cloud storage is safe here.";
+  public static final String UNENCRYPTED_MESSAGE = "Make sure using unencrypted cloud storage is safe here.";
   public static final String FORMAT_OMITTING = "Omitting %s enables clear-text storage. Make sure it is safe here.";
 
   @Override
-  public void initialize(InitContext init) {
-    new AzureUnencryptedCloudServicesCheckPart().initialize(init);
+  protected void registerResourceConsumer() {
+    register("azurerm_data_lake_store",
+      resource -> resource.attribute("encryption_state")
+        .reportIfValueMatches("Disabled", UNENCRYPTED_MESSAGE));
+
+    register("azurerm_managed_disk",
+      resource -> resource.attribute("disk_encryption_set_id")
+        .reportIfAbsence(FORMAT_OMITTING));
+
+    register("azurerm_mysql_server",
+      resource -> resource.attribute("infrastructure_encryption_enabled")
+        .reportIfAbsence(FORMAT_OMITTING)
+        .reportIfFalse(UNENCRYPTED_MESSAGE));
   }
 }
